+++ src/main/java/com/nucleus/customerservice/loandisbursal/database/LoanDisbursalDAO.java
@@ -1,28 +1,29 @@
 package com.nucleus.customerservice.loandisbursal.database;
 
-import com.nucleus.customerservice.loandisbursal.model.Customer;
-import com.nucleus.customerservice.loandisbursal.model.LoanApplication;
+import com.nucleus.customer.model.Customer;
+import com.nucleus.loanaplications.model.LoanApplications;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.util.Set;
+import java.util.List;
 
 @Component
 public class LoanDisbursalDAO {
     @Autowired
     private SessionFactory sessionFactory;
 
-    public LoanApplication getLoanDetails(int loanApplicationNumber){
+    public LoanApplications getLoanDetails(int loanApplicationNumber){
         Session session = sessionFactory.openSession();
-        LoanApplication loanApplication = session.get(LoanApplication.class, loanApplicationNumber);
+        LoanApplications loanApplication = session.get(LoanApplications.class, loanApplicationNumber);
         return loanApplication;
     }
 
-    public Set<LoanApplication> getCustomerLoanDetails(String customerCode){
+    public List<LoanApplications> getCustomerLoanDetails(String customerCode){
         Session session = sessionFactory.openSession();
         Customer customer = session.get(Customer.class, customerCode);
-        return customer.getLoanApplications();
+        return customer.getAllloanApplications();
     }
 }
