+++ src/main/java/com/nucleus/customerservice/loandisbursal/service/LoanDisbursalService.java
@@ -1,21 +1,21 @@
 package com.nucleus.customerservice.loandisbursal.service;
 
 import com.nucleus.customerservice.loandisbursal.database.LoanDisbursalDAO;
-import com.nucleus.customerservice.loandisbursal.model.LoanApplication;
+import com.nucleus.loanaplications.model.LoanApplications;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Set;
+import java.util.List;
 @Service
 public class LoanDisbursalService {
     @Autowired
     private LoanDisbursalDAO loanDisbursalDao;
 
-    public LoanApplication getLoanDetails(int loanApplicationNumber){
+    public LoanApplications getLoanDetails(int loanApplicationNumber){
         return loanDisbursalDao.getLoanDetails(loanApplicationNumber);
     }
 
-    public Set<LoanApplication> getCustomerLoanDetails(String customerCode){
+    public List<LoanApplications> getCustomerLoanDetails(String customerCode){
         return loanDisbursalDao.getCustomerLoanDetails(customerCode);
     }
 }
