+++ src/main/java/com/nucleus/loanaplications/dao/LoanApplicationDaoInterface.java
@@ -1,10 +1,10 @@
 package com.nucleus.loanaplications.dao;
 
-import com.nucleus.customerservice.loandisbursal.model.LoanApplication;
+import com.nucleus.loanaplications.model.LoanApplications;
 
 import java.util.List;
 
 public interface LoanApplicationDaoInterface {
-    boolean addApplication(LoanApplication loanApplication);
-    List<LoanApplication> getLoanApplicationList();
+    boolean addApplication(LoanApplications loanApplication);
+    List<LoanApplications> getLoanApplicationList();
 }
