+++ src/main/webapp/WEB-INF/views/loan disbursal/ErrorPage.jsp
@@ -6,8 +6,31 @@
 <html>
 <head>
     <title>Error</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet"
+          href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.2/css/bootstrap.css">
+    <link rel="stylesheet"
+          href="https://cdn.datatables.net/1.10.22/css/jquery.dataTables.min.css">
+    <script
+            src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <script src="https://code.jquery.com/jquery-3.5.1.js"></script>
+    <script
+            src="https://cdn.datatables.net/1.10.22/js/jquery.dataTables.min.js"></script>
+    <script
+            src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
+    <script
+            src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 </head>
 <body>
-<c:out value="${Message}" />
+<div class="container-fluid">
+    <!-- Section Heading -->
+    <div class="row pt-3 pl-3 flex-column">
+        <h2 class="  display-3" style="font-size: 30px">
+             <c:out value="${Message}" />
+        </h2>
+    </div>
+    <hr width="" color="#b3b3b3">
+</div>
+
 </body>
 </html>
